Relocate view models
