+++ app/src/main/java/com/hegocre/nextcloudpasswords/ui/components/NCPApp.kt
@@ -48,9 +48,9 @@ import androidx.navigation.compose.rememberNavController
 import com.hegocre.nextcloudpasswords.R
 import com.hegocre.nextcloudpasswords.api.FoldersApi
 import com.hegocre.nextcloudpasswords.data.password.Password
-import com.hegocre.nextcloudpasswords.data.viewmodels.PasswordsViewModel
 import com.hegocre.nextcloudpasswords.ui.NCPScreen
 import com.hegocre.nextcloudpasswords.ui.theme.NextcloudPasswordsTheme
+import com.hegocre.nextcloudpasswords.ui.viewmodels.PasswordsViewModel
 import kotlinx.coroutines.launch
 
 @OptIn(
